git-report: skip failing reports outside worktrees
Not all aspects of the git-report script make sense outside a Git
worktree, so test if we're in a Git worktree before performing them.

This handles the case where the command is run from inside a bare
repository; it doesn't handle the case where the command is run from
something that isn't a Git repository at all: that's covered by the
existing `git rev-parse --git-dir` returning an error code and the `set
-e` meaning the script will immediately bail out in the face of an
unhandled error.