refresh-git: add script for rebuilding Git
I want to rebuild and reinstall Git from source sufficiently frequently
that it makes sense to have a script for it, in the name of avoiding
needing to remember the magic commands myself.

Currently this is only useful to me on Linux systems: on macOS I can use
Homebrew to get the latest version of Git, and on Cygwin I'm using the
official release build in the name of dogfood testing.