cygwin-setup: use `cygstart` to start setup
If escalation to admin permissions isn't necessary, launching the setup
program directly means it's a sub-process of the Bash shell.  That means
when the setup program wants to update Bash or cygwin1.dll or any other
dependency, it can't kill Bash to do the install without killing itself.
To work around this, use cygstart to avoid the setup program being a
sub-process.