cygwin-setup: only update when necessary
When downloading a new version of the Cygwin setup program, if possible
only download if the version on the server is actually newer than the
extant local one.

This requires changing the default from wget to curl, since wget doesn't
have the option to download only if the server version is newer.