ssh-agent: handle systems sharing home dir
Sometimes when using Cygwin, I want to have both 32-bit and 64-bit
Cygwin installed on the same machine and using the same home directory.
That means my ssh-agent handling code needs to be able to cope with
separate ssh-agent instances running in the 32- and 64-bit environments,
with different PIDs.  To allow that to work, if $HOSTTYPE is defined,
use that in the file name storing the relevant ssh-agent PID.