git-config: use https for GitHub pulls by default
Always use SSH for git pushes to GitHub, because that's still much
easier than setting up the HTTPS authentication.  But for pulls, most of
the time I don't need authentication, and requiring authentication makes
automatic background pulls (such as provided by bash-git-prompt) more
awkward, particularly now that I'm not reliably running ssh-add on
login, so use HTTPS for those.

For the few repositories that genuinely require SSH access to pull, it's
still possible to explicitly give the repository URL as an SSH one, and
that won't be overrided by this configuration.