systemd ssh-agent: reconfigure
I do want systemd to provide an SSH agent when I log in.  I want this so
much, in fact, that I want it to happen even if I've never installed X11
on the system in question.  So override the default ssh-agent.service
file that only works if X11 packages have been installed and not purged,
with one that works regardless provided you know where to look for the
socket.

This might need re-wrangling if and when I want to use this config
*with* a system with X11, but I'll burn that bridge when I come to it...