systemd: move config to separate repo
It seems the systemd user service doesn't cope well with having files
symlinked -- the disable/enable commands don't work the same as if the
unit files were simple files -- but can cope just fine with the config
folder being a symlink.  So move the systemd config into a separate repo
and create a symlink to that repo.