rm-final-eol: new tool for removing eols
It seems `systemctl edit` and the like prefer to leave files without a
final line ending character.  Which means Git diff logs and the like end
up with flapping about between presence and absence of final EOLs
depending on whether I used the systemctl tools or Vim to edit the
files.  Provide a script to tidy that up.