ssh-config: add ro.github.com
Some repositories on GitHub require authentication to access them even
for read-only purposes.  I want to be able to automate read-only access
to them, though, so to allow that, I can create a repository-specific
read-only deploy key, and store that without password protection.

However these keys are repository-specific, so it makes sense to keep
them alongside the repositories.  Plus I want to make sure that for
write operations, I still pick up the normal key, as otherwise SSH will
successfully log in with the deploy key then a `git push` or similar
will fail.  To get this to work, set up the ssh config so that if a host
is specified as `ro.github.com`, it's actually talking about
`github.com`, but SSH should instead use the deploy key and ignore any
other keys.