ssh-config: allow overriding defaults
ssh will use the first value it encounters in a config file, so values
at the top of the file will always override ones specified in Host/Match
blocks later, including in later Include directives.  If I'm specifying
something in a Match or Host block, it's because I want it to override
whatever my default config is, so move the default config to a final
Host block at the end of the config file.