ssh-config: use legacy keyword
In commit V_8_4_P1-166-gee9c0da8 in the OpenSSH repository, the project
renamed the apparently-misleading PubkeyAcceptedKeyTypes config key to
PubkeyAcceptedAlgorithms.  Which is all well and good for Cygwin's
up-to-date OpenSSH, but no good at all for the version of OpenSSH
available on Debian stable.  The old version is retained as an alias, so
use the old version.