homeshick-report: improve day-to-day handling
If using homeshick-report on a system that doesn't do homeshick
monitoring using systemd, wrap the homeshick-report executable with a
function which will also update the last-run flagfile when run to check
all homeshick repos.

However, only update the flag file if the report was actually that
everything was up-to-date.  If it wasn't, it should be hassling me more
frequently!