homeshick-report: handle like cygwin_setup &c
The homeshick-report command can be slow, or take forever in the case
where a password is needed to access a remote repo.  Don't run it
automatically when starting Bash, and instead just warn if it has been a
while since it was last run, following the same paradigm as for
cygwin_setup.