ssh: move shared known_hosts file
I'd originally planned to use known_hosts to track things in my
Homeshick castle, and known_hosts2 to track things that were going to be
local to the system.  That's suboptimal for two reasons:

- It appears known_hosts2 is a legacy file from the days when SSH was
  first adopting ssh version 2, so using it for this purpose seemed like
  the sort of fragile thing that might cause problems in future.
- When OpenSSH updated the known_hosts file, it did so my moving the
  original to known_hosts.old, and creating a new file.  That broke
  Homeshick's symlinking.

Instead, specify explicit UserKnownHosts files, with the first one being
the traditional ~/.ssh/known_hosts that OpenSSH can update freely, and
the second being ~/.ssh/known_hosts_shared, which will be tracked and
shared in this Homeshick castle, and which -- because it's not the first
file in the list -- OpenSSH (hopefully) won't try to update.