bashrc: remove `cd` for WSL
It looks like the not-starting-in-home-directory behaviour I was seeing
was (a) specific to WSLtty, and (b) specific to WSLtty when started
without the -~ option.  Given it's easy enough to add that option, and
some WSLtty shortcuts come with it already specified, I'm removing the
`cd` from my .bashrc, since it removes the idempotence that .bashrc
otherwise (mostly?) has.