taskloop: check for closed children before drawing
At the start of the loop, as well as checking what child processes have
assigned PIDs, also check that the process still exists.  In particular,
this means when a task command has changed the waited-for files, so both
readpid and toilpid will exit, that doesn't cause the table to be
redrawn twice in quick succession.