taskloop: try adding a delay for "immediate" files
If a file that should trigger an immediate change is changed from within
the interactive Bash session, there's a race between toil and Bash
exiting.  If Bash exits first, that can cause the taskloop to loop
twice, once because Bash exited, and a second time because toil exits
after the check for whether toil has already exited.